requirements.txt
streamlit
google-generativeai
python-dotenv
langchain  #for ollama
gradio   #for frontend